requirements.txt
Django==4.2.7
requests==2.31.0
slumber==0.7.1
python-dotenv==1.0.0
